requirements.txt
numpy==1.13.3
keras==2.2.4
pandas==0.24.2
matplotlib==2.0.2
scikit-learn==0.20.1
h5py==2.9.0
pillow==3.4.2
